feat: add conversao tempo

+++ prj_VisualValue/Frm_editar_custos.vb
@@ -159,20 +159,22 @@
     Private Sub dgv_listaperfil_CellContentClick(sender As Object, e As DataGridViewCellEventArgs) Handles dgv_listaperfil.CellContentClick
         With dgv_listaperfil
             Try
-                If .CurrentRow.Cells(3).Selected Then
-                    resp = MsgBox("Deseja apagar o perfil '" & .CurrentRow.Cells(1).Value & "'?",
-                        MsgBoxStyle.Question + MsgBoxStyle.YesNo, "AVISO")
-                    If resp = MsgBoxResult.Yes Then
-                        aux_id_perfil = .CurrentRow.Cells(0).Value
+                If .RowCount > 0 Then
+                    If .CurrentRow.Cells(3).Selected Then
+                        resp = MsgBox("Deseja apagar o perfil '" & .CurrentRow.Cells(1).Value & "'?",
+                            MsgBoxStyle.Question + MsgBoxStyle.YesNo, "AVISO")
+                        If resp = MsgBoxResult.Yes Then
+                            aux_id_perfil = .CurrentRow.Cells(0).Value
 
-                        sql = "DELETE FROM tb_perfil_custos WHERE id_perfil_custos = " & aux_id_perfil
-                        tabela = banco.Execute(sql)
+                            sql = "DELETE FROM tb_perfil_custos WHERE id_perfil_custos = " & aux_id_perfil
+                            tabela = banco.Execute(sql)
 
-                        sql = "DELETE FROM tb_custos_mensais WHERE id_perfil_custos = " & aux_id_perfil
-                        tabela = banco.Execute(sql)
+                            sql = "DELETE FROM tb_custos_mensais WHERE id_perfil_custos = " & aux_id_perfil
+                            tabela = banco.Execute(sql)
 
-                        MsgBox("Perfil de custos mensais deletado com sucesso!", MsgBoxStyle.Information + MsgBoxStyle.OkOnly, "AVISO")
-                        carregar_custos_mensais()
+                            MsgBox("Perfil de custos mensais deletado com sucesso!", MsgBoxStyle.Information + MsgBoxStyle.OkOnly, "AVISO")
+                            carregar_custos_mensais()
+                        End If
                     End If
                 End If
             Catch ex As Exception
@@ -180,9 +182,11 @@
             End Try
 
             Try
-                If .CurrentRow.Cells(2).Selected Then
-                    aux_id_perfil = .CurrentRow.Cells(0).Value
-                    alt_modo_edicao_custos_mensais(aux_id_perfil, True)
+                If .RowCount > 0 Then
+                    If .CurrentRow.Cells(2).Selected Then
+                        aux_id_perfil = .CurrentRow.Cells(0).Value
+                        alt_modo_edicao_custos_mensais(aux_id_perfil, True)
+                    End If
                 End If
             Catch ex As Exception
                 MsgBox("Erro ao entrar no modo de edição")
